boards: move MonotonicWrapper and DummyNfc up to soc
