+++ components/boards/src/lib.rs
@@ -7,6 +7,7 @@ use cortex_m_rt::ExceptionFrame;
 
 pub mod flash;
 pub mod init;
+pub mod nfc;
 pub mod runtime;
 pub mod soc;
 pub mod store;
