+++ components/boards/src/nfc.rs
@@ -1,0 +1,15 @@
+use nfc_device::traits::nfc::{Device as NfcDevice, Error as NfcError, State as NfcState};
+
+pub struct DummyNfc;
+
+impl NfcDevice for DummyNfc {
+    fn read(&mut self, _buf: &mut [u8]) -> Result<NfcState, NfcError> {
+        Err(NfcError::NoActivity)
+    }
+    fn send(&mut self, _buf: &[u8]) -> Result<(), NfcError> {
+        Err(NfcError::NoActivity)
+    }
+    fn frame_size(&self) -> usize {
+        0
+    }
+}
