+++ components/boards/src/soc/lpc55.rs
@@ -13,7 +13,6 @@ use lpc55_hal::{
 };
 
 pub mod clock_controller;
-pub mod monotonic;
 
 type UsbPeripheral = lpc55_hal::peripherals::usbhs::EnabledUsbhsDevice;
